+++ include/stars-electrostatics.h
@@ -1,10 +1,11 @@
 typedef struct STARS_esdata
 {
-    int count;
+    size_t count;
     double *point;
 } STARS_esdata;
 
-int STARS_esdata_block_kernel(int nrows, int ncols, int *irow, int *icol,
-        void *row_data, void *col_data, void *result);
-STARS_esdata *STARS_gen_esdata(int row_blocks, int col_blocks, int block_size);
+int STARS_esdata_block_kernel(size_t nrows, size_t ncols, size_t *irow,
+        size_t *icol, void *row_data, void *col_data, void *result);
+STARS_esdata *STARS_gen_esdata(size_t row_blocks, size_t col_blocks,
+        size_t block_size);
 void STARS_esdata_free(STARS_esdata *data);
